Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -43,7 +43,7 @@ def get_average_age_for_season(data, season)
   total = 0
   num = 0
   data[season].each do |a|
-      a = a["age"].to_i
+      a = a["age"].to_f
       total+= a
       num+=1
   end
